use props instead of hard-coded placeholder
I had problems with using a placeholder in my app. First I've tried by using an additional property "placeholder" but for unknown reasons the original "Tell a Story" didn't want to disappear. So I introduced a new property "placeholder". If set to null/undefined the original "Tell a Story" will be used instead.

I hope this solution is OK.